collection, add, update, gson V4.1

+++ main/java/ua/goit/moneybot/user/UserInformation.java
@@ -2,13 +2,13 @@ package ua.goit.moneybot.user;
 
 public class UserInformation {
     String name = "User";
-    int chatId = 1;
+    long chatId = 1;
     UserSettings userSettings = new UserSettings();
 
     public UserInformation() {
     }
 
-    public UserInformation(String name, int chatId, UserSettings userSettings) {
+    public UserInformation(String name, long chatId, UserSettings userSettings) {
         this.name = name;
         this.chatId = chatId;
         this.userSettings = userSettings;
